[SAFE-112] User deletion working

+++ schema/mutation/deleteUsers.ts
@@ -1,14 +1,13 @@
-import { GraphQLNonNull, GraphQLID, GraphQLError, GraphQLList } from "graphql";
+import { GraphQLNonNull, GraphQLID, GraphQLError, GraphQLList, GraphQLBoolean } from "graphql";
 import errors from "../../const/errors";
-import { Role, User } from "../../models";
+import { User } from "../../models";
 import { AppAbility } from "../../security/defineAbilityFor";
-import { RoleType, UserType } from "../types";
 
 export default {
     /*  Deletes a user.
         Throws an error if not logged or authorized.
     */
-    type: UserType,
+    type: GraphQLBoolean,
     args: {
         ids: { type: new GraphQLNonNull(new GraphQLList(GraphQLID)) }
     },
@@ -16,7 +15,12 @@ export default {
         // Authentication check
         const user = context.user;
         if (!user) { throw new GraphQLError(errors.userNotLogged); }
-
-        await User.deleteMany({_id: {$in: args.ids}});
+        const ability: AppAbility = user.ability;
+        if (ability.can('delete', 'User')) {
+            await User.deleteMany({_id: {$in: args.ids}});
+            return true;
+        } else {
+            throw new GraphQLError(errors.permissionNotGranted);
+        }
     }
 }
